backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
langchain>=0.0.267
langchain-openai>=0.0.5
beautifulsoup4>=4.12.2
requests>=2.31.0
chromadb>=0.4.18
supabase>=1.0.3
python-dotenv>=1.0.0
